Reworked array and sorting
Changed the array to a non associative array as it caused limitations and made things more difficult. Now the process of making the array is more streamlined and sturdy. Added in a sorting function that will sort the array of events upon finishing building it. Sorts based upon start time and does take into account am vs pm.