Basic Form Addition
Added a very basic form that only allows adding one event. It will then add it to the CSV file and update the schedule. Javascript will be helpful to allow a button to add a new line in the form with the same fields allowing for multiple events for a single tutor to be added. Additionally will need to improve code to allow multiple events to be submitted at once.