Overhaul calendar building process
I changed the process of building the calendar to be 3 seperate functions each capable of being individually included into any php file. Now it only reads from a temporary CSV to allow for changes to the CSV to be made by the admin, preview the changes, and then decide to discard all progress without impacting the primary CSV. This also means a temporary tutoring page is made which the admin can view to see a preview of what people viewing the tutoring page will see. This again allows for any changes they made to be discarded without impacting the primary tutoring page. Lastly, I added a finalizeTutoingPage.php file that reads the temp csv and tutoring page and writes them onto the primary csv and tutoring page making the changes final. This means all progress up until the last step wont be pushed through to the students tutoring page until the finalize file is run.