Addition of form and edit tutor functionality
Added more functions to handle the edit tutor form. Functions build the form, both blank and pre filled. Buttons exist for adding new lines to the form, but lack the java script needed to actually add the lines. The forms can be submitted and the changes will be saved. Have not fully tested changing names of pre existing tutors, but that functionality should be working. Added a return button to the edit tutor page to allow returning to the home admin page. Changed buttons on main admin page into hidden forms with the buttons being the submit button. They have predefined values and are hidden and thus unchangeable allowing for necessary information to be pased onto the next page. Still need to add in a button on main page to push changes to public accessible tutoring page and still need to actually write the function that builds that page. Will hold off and doing that part until tomorrow as it is simple and fairly straight forward. Will likely use the same concept of a form as a button. The edit tutor page lacks any style but can be very easily adapted to style just as with the main page. Also need a button to delete events and tutors, but the functions to delete the tutor do exist, just have not been implemented fully.