Added the tutoring page
I implemented the publish changes functiniality. It now brings you to a new page via a link and that page writes just the schedule to an html file. The page is then redirected back to the admin page and shows a confirmation message to the user. The tutoring page for public use is php but it draws its schedule by including the schedule.html file which is only every changed when the publish changes button is pressed.